+++ src/pre_commit_hooks/ast_checks/__init__.py
@@ -7,6 +7,7 @@ from .meaningless_vars import MeaninglessVarsCheck
 from .misplaced_comment import MisplacedCommentCheck
 from .redundant_assignment import RedundantAssignmentCheck
 from .redundant_dict_get import RedundantDictGetCheck
+from .redundant_enum_value import RedundantEnumValueCheck
 from .redundant_super_init import RedundantSuperInitCheck
 from .redundant_type_conversion import RedundantTypeConversionCheck
 from .unused_pytriage import UnusedPytriageCheck
@@ -25,4 +26,5 @@ ALL_CHECKS: list[type[ASTCheck]] = [
     RedundantTypeConversionCheck,
     UnusedPytriageCheck,
     RedundantDictGetCheck,
+    RedundantEnumValueCheck,
 ]